refactor getPreFilledLink
